+++ test/equatable_test.dart
@@ -4,73 +4,60 @@
 import 'dart:convert';
 
 import 'package:equatable/equatable.dart';
-import 'package:equatable/src/equatable_utils.dart';
 import 'package:test/test.dart';
 
 import 'custom_list.dart';
 
 class NonEquatable {}
 
-class EmptyEquatable extends Equatable {
+@Equatable()
+class EmptyEquatable {
   const EmptyEquatable();
-
-  @override
-  List<Object> get props => [];
 }
 
-class SimpleEquatable<T extends Object> extends Equatable {
+@Equatable()
+class SimpleEquatable<T extends Object> {
   const SimpleEquatable(this.data);
-
   final T data;
-
-  @override
-  List<Object> get props => [data];
 }
 
-class MultipartEquatable<T extends Object> extends Equatable {
+@Equatable()
+class MultipartEquatable<T extends Object> {
   MultipartEquatable(this.d1, this.d2);
 
   final T d1;
   final T d2;
-
-  @override
-  List<Object> get props => [d1, d2];
 }
 
-class OtherEquatable extends Equatable {
+@Equatable()
+class OtherEquatable {
   const OtherEquatable(this.data);
 
   final String data;
-
-  @override
-  List<Object> get props => [data];
 }
 
 enum Color { blonde, black, brown }
 
-class ComplexEquatable extends Equatable {
+@Equatable()
+class ComplexEquatable {
   const ComplexEquatable({this.name, this.age, this.hairColor, this.children});
 
   final String? name;
   final int? age;
   final Color? hairColor;
   final List<String>? children;
-
-  @override
-  List<Object?> get props => [name, age, hairColor, children];
 }
 
-class EquatableData extends Equatable {
+@Equatable()
+class EquatableData {
   const EquatableData({required this.key, required this.value});
 
   final String key;
   final Object value;
-
-  @override
-  List<Object> get props => [key, value];
 }
 
-class Credentials extends Equatable {
+@Equatable()
+class Credentials {
   const Credentials({required this.username, required this.password});
 
   factory Credentials.fromJson(Map<String, dynamic> json) {
@@ -89,84 +76,19 @@ class Credentials extends Equatable {
       'password': password,
     };
   }
-
-  @override
-  List<Object> get props => [username, password];
-}
-
-class ComplexStringify extends Equatable {
-  ComplexStringify({this.name, this.age, this.hairColor});
-
-  final String? name;
-  final int? age;
-  final Color? hairColor;
-
-  @override
-  List<Object?> get props => [name, age, hairColor];
-
-  @override
-  bool get stringify => true;
-}
-
-class SuperLongPropertiesStringify extends Equatable {
-  SuperLongPropertiesStringify(this.a, this.b, this.c, this.d, this.e, this.f);
-
-  final String a;
-  final String b;
-  final String c;
-  final String d;
-  final String e;
-  final String f;
-
-  @override
-  List<Object> get props => [a, b, c, d, e, f];
-
-  @override
-  bool get stringify => true;
-}
-
-class ExplicitStringifyFalse extends Equatable {
-  ExplicitStringifyFalse({this.name, this.age, this.hairColor});
-
-  final String? name;
-  final int? age;
-  final Color? hairColor;
-
-  @override
-  List<Object?> get props => [name, age, hairColor];
-
-  @override
-  bool get stringify => false;
 }
 
 void main() {
-  late bool globalStringify;
-
-  setUp(() {
-    globalStringify = EquatableConfig.stringify;
-  });
-
-  tearDown(() {
-    EquatableConfig.stringify = globalStringify;
-  });
-
   group('Empty Equatable', () {
-    test('should correct toString', () {
-      final instance = EmptyEquatable();
-      expect(instance.toString(), 'EmptyEquatable()');
-    });
-
     test('should return true when instance is the same', () {
       final instance = EmptyEquatable();
       expect(instance == instance, true);
     });
 
     test('should return correct hashCode', () {
-      final instance = EmptyEquatable();
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
-      );
+      final instanceA = EmptyEquatable();
+      final instanceB = EmptyEquatable();
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
     test('should return true when instances are different', () {
@@ -184,33 +106,15 @@ void main() {
   });
 
   group('Simple Equatable (string)', () {
-    test('should correct toString', () {
-      final instance = SimpleEquatable('simple');
-      expect(instance.toString(), 'SimpleEquatable<String>(simple)');
-    });
-
-    test('should correct toString when EquatableConfig.stringify is false', () {
-      EquatableConfig.stringify = false;
-      final instance = SimpleEquatable('simple');
-      expect(instance.toString(), 'SimpleEquatable<String>');
-    });
-
     test('should return true when instance is the same', () {
       final instance = SimpleEquatable('simple');
       expect(instance == instance, true);
     });
 
     test('should return correct hashCode', () {
-      final instance = SimpleEquatable('simple');
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
-      );
-    });
-
-    test('should return correct toString', () {
-      final instance = SimpleEquatable('simple');
-      expect(instance.toString(), 'SimpleEquatable<String>(simple)');
+      final instanceA = SimpleEquatable('simple');
+      final instanceB = SimpleEquatable('simple');
+      expect(instanceA.hashCode, instanceB.hashCode);
     });
 
     test('should return true when instances are different', () {
@@ -240,22 +144,15 @@ void main() {
   });
 
   group('Simple Equatable (number)', () {
-    test('should return correct toString', () {
-      final instance = SimpleEquatable(0);
-      expect(instance.toString(), 'SimpleEquatable<int>(0)');
-    });
-
     test('should return true when instance is the same', () {
       final instance = SimpleEquatable(0);
       expect(instance == instance, true);
     });
 
     test('should return correct hashCode', () {
-      final instance = SimpleEquatable(0);
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
-      );
+      final instanceA = SimpleEquatable(0);
+      final instanceB = SimpleEquatable(0);
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
     test('should return true when instances are different', () {
@@ -279,22 +176,15 @@ void main() {
   });
 
   group('Simple Equatable (bool)', () {
-    test('should correct toString', () {
-      final instance = SimpleEquatable(true);
-      expect(instance.toString(), 'SimpleEquatable<bool>(true)');
-    });
-
     test('should return true when instance is the same', () {
       final instance = SimpleEquatable(true);
       expect(instance == instance, true);
     });
 
     test('should return correct hashCode', () {
-      final instance = SimpleEquatable(true);
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
-      );
+      final instanceA = SimpleEquatable(true);
+      final instanceB = SimpleEquatable(true);
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
     test('should return true when instances are different', () {
@@ -318,22 +208,9 @@ void main() {
   });
 
   group('Simple Equatable (map)', () {
-    test('should correct toString', () {
-      final instance = SimpleEquatable(<String, dynamic>{});
-      expect(instance.toString(), 'SimpleEquatable<Map<String, dynamic>>({})');
-    });
-
     test('should return true when instance is the same', () {
       final instance = SimpleEquatable({'a': 1, 'b': 2, 'c': 3});
       expect(instance == instance, true);
-    });
-
-    test('should return correct hashCode', () {
-      final instance = SimpleEquatable({'a': 1, 'b': 2, 'c': 3});
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
-      );
     });
 
     test('should have same hashCode when values are equal', () {
@@ -364,14 +241,6 @@ void main() {
   });
 
   group('Simple Equatable (map custom key)', () {
-    test('should correct toString', () {
-      final instance = SimpleEquatable(<SimpleEquatable<String>, dynamic>{});
-      expect(
-        instance.toString(),
-        'SimpleEquatable<Map<SimpleEquatable<String>, dynamic>>({})',
-      );
-    });
-
     test('should return true when instance is the same', () {
       final instance = SimpleEquatable(
         {
@@ -384,17 +253,21 @@ void main() {
     });
 
     test('should return correct hashCode', () {
-      final instance = SimpleEquatable(
+      final instanceA = SimpleEquatable(
         {
           SimpleEquatable<String>('a'): 1,
           SimpleEquatable<String>('b'): 2,
           SimpleEquatable<String>('c'): 3,
         },
       );
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
+      final instanceB = SimpleEquatable(
+        {
+          SimpleEquatable<String>('a'): 1,
+          SimpleEquatable<String>('b'): 2,
+          SimpleEquatable<String>('c'): 3,
+        },
       );
+      expect(instanceA.hashCode, instanceB.hashCode);
     });
 
     test('should have same hashCode when values are equal', () {
@@ -467,18 +340,6 @@ void main() {
   });
 
   group('Simple Equatable (Equatable)', () {
-    test('should correct toString', () {
-      final instance = SimpleEquatable(
-        EquatableData(
-          key: 'foo',
-          value: 'bar',
-        ),
-      );
-      expect(
-        instance.toString(),
-        'SimpleEquatable<EquatableData>(EquatableData(foo, bar))',
-      );
-    });
     test('should return true when instance is the same', () {
       final instance = SimpleEquatable(
         EquatableData(
@@ -490,16 +351,19 @@ void main() {
     });
 
     test('should return correct hashCode', () {
-      final instance = SimpleEquatable(
+      final instanceA = SimpleEquatable(
         EquatableData(
           key: 'foo',
           value: 'bar',
         ),
       );
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
+      final instanceB = SimpleEquatable(
+        EquatableData(
+          key: 'foo',
+          value: 'bar',
+        ),
       );
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
     test('should return true when instances are different', () {
@@ -548,26 +412,20 @@ void main() {
   });
 
   group('Multipart Equatable', () {
-    test('should correct toString', () {
-      final instance = MultipartEquatable('s1', 's2');
-      expect(instance.toString(), 'MultipartEquatable<String>(s1, s2)');
-    });
-
     test('should return true when instance is the same', () {
       final instance = MultipartEquatable('s1', 's2');
       expect(instance == instance, true);
     });
 
     test('should return correct hashCode', () {
-      final instance = MultipartEquatable('s1', 's2');
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
-      );
+      final instanceA = MultipartEquatable('s1', 's2');
+      final instanceB = MultipartEquatable('s1', 's2');
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
-    test('should return different hashCodes when property order has changed',
-        () {
+    test(
+        'should return different hashCodes '
+        'when property order has changed', () {
       final instance1 = MultipartEquatable('s1', 's2');
       final instance2 = MultipartEquatable('s2', 's1');
       expect(instance1.hashCode == instance2.hashCode, isFalse);
@@ -598,18 +456,6 @@ void main() {
   });
 
   group('Complex Equatable', () {
-    test('should correct toString', () {
-      final instance = ComplexEquatable(
-        name: 'Joe',
-        age: 40,
-        hairColor: Color.black,
-        children: ['Bob'],
-      );
-      expect(
-        instance.toString(),
-        'ComplexEquatable(Joe, 40, Color.black, [Bob])',
-      );
-    });
     test('should return true when instance is the same', () {
       final instance = ComplexEquatable(
         name: 'Joe',
@@ -621,16 +467,19 @@ void main() {
     });
 
     test('should return correct hashCode', () {
-      final instance = ComplexEquatable(
+      final instanceA = ComplexEquatable(
         name: 'Joe',
         age: 40,
         hairColor: Color.black,
         children: ['Bob'],
       );
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
+      final instanceB = ComplexEquatable(
+        name: 'Joe',
+        age: 40,
+        hairColor: Color.black,
+        children: ['Bob'],
       );
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
     test('should return true when instances are different', () {
@@ -711,20 +560,6 @@ void main() {
   });
 
   group('Json Equatable', () {
-    test('should correct toString', () {
-      final instance = Credentials.fromJson(
-        json.decode(
-          '''
-        {
-          "username":"Admin",
-          "password":"admin"
-        }
-        ''',
-        ) as Map<String, dynamic>,
-      );
-      expect(instance.toString(), 'Credentials(Admin, admin)');
-    });
-
     test('should return true when instance is the same', () {
       final instance = Credentials.fromJson(
         json.decode(
@@ -740,7 +575,7 @@ void main() {
     });
 
     test('should return correct hashCode', () {
-      final instance = Credentials.fromJson(
+      final instanceA = Credentials.fromJson(
         json.decode(
           '''
         {
@@ -750,10 +585,17 @@ void main() {
         ''',
         ) as Map<String, dynamic>,
       );
-      expect(
-        instance.hashCode,
-        instance.runtimeType.hashCode ^ mapPropsToHashCode(instance.props),
+      final instanceB = Credentials.fromJson(
+        json.decode(
+          '''
+        {
+          "username":"Admin",
+          "password":"admin"
+        }
+        ''',
+        ) as Map<String, dynamic>,
       );
+      expect(instanceA.hashCode, equals(instanceB.hashCode));
     });
 
     test('should return true when instances are different', () {
@@ -1014,62 +856,4 @@ void main() {
       });
     });
   });
-
-  group('To String Equatable', () {
-    test('with Complex stringify', () {
-      final instanceA = ComplexStringify();
-      final instanceB = ComplexStringify(name: 'Bob', hairColor: Color.black);
-      final instanceC = ComplexStringify(
-        name: 'Joe',
-        age: 50,
-        hairColor: Color.blonde,
-      );
-      expect(instanceA.toString(), 'ComplexStringify(null, null, null)');
-      expect(instanceB.toString(), 'ComplexStringify(Bob, null, Color.black)');
-      expect(instanceC.toString(), 'ComplexStringify(Joe, 50, Color.blonde)');
-    });
-
-    test('with SuperLongProperties stringify', () {
-      final instance = SuperLongPropertiesStringify(
-        'aaaaaaaaaaaaaaa',
-        'aaaaaaaaaaaaaaa',
-        'aaaaaaaaaaaaaaa',
-        'aaaaaaaaaaaaaaa',
-        'aaaaaaaaaaaaaaa',
-        'aaaaaaaaaaaaaaa',
-      );
-      expect(
-        instance.toString(),
-        'SuperLongPropertiesStringify(aaaaaaaaaaaaaaa, aaaaaaaaaaaaaaa, '
-        'aaaaaaaaaaaaaaa, aaaaaaaaaaaaaaa, aaaaaaaaaaaaaaa, '
-        'aaaaaaaaaaaaaaa)',
-      );
-    });
-
-    test('with ExplicitStringifyFalse stringify', () {
-      final instanceA = ExplicitStringifyFalse();
-      final instanceB = ExplicitStringifyFalse(
-        name: 'Bob',
-        hairColor: Color.black,
-      );
-      final instanceC = ExplicitStringifyFalse(
-        name: 'Joe',
-        age: 50,
-        hairColor: Color.blonde,
-      );
-      expect(instanceA.toString(), 'ExplicitStringifyFalse');
-      expect(instanceB.toString(), 'ExplicitStringifyFalse');
-      expect(instanceC.toString(), 'ExplicitStringifyFalse');
-    });
-  });
-}
-
-// test that subclasses of `Equatable` can have const constructors
-class ConstTest extends Equatable {
-  const ConstTest(this.a);
-
-  final int a;
-
-  @override
-  List<Object> get props => [a];
 }