+++ test/main.dart
@@ -1,0 +1,5 @@
+import 'equatable_test.dart' as equatable_test;
+
+void main() {
+  equatable_test.main();
+}